requirements.txt
alembic==1.14.1
autoflake==2.3.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask>=3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-mongoengine==1.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mongoengine==0.29.1
packaging==24.2
pyflakes==3.2.0
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
